Switch to @gustavnikolaj/async-main-wrap
