+++ package.json
@@ -15,6 +15,7 @@
   "author": "Andreas Lind <andreaslindpetersen@gmail.com>",
   "license": "BSD-3-Clause",
   "dependencies": {
+    "@gustavnikolaj/async-main-wrap": "^2.0.0",
     "assetgraph": "^5.1.0",
     "error-stack-parser": "^2.0.1",
     "get-stdin": "^6.0.0",
